feat 1.2: comment-view.dto.ts

+++ src/modules/bloggers-platform/posts/application/use-cases/delete-post.usecase.ts
@@ -1,6 +1,6 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import PostsRepository from '../../infrastructure/posts.repository';
-import { DomainException } from '@core/exceptions/filters/domain-exceptions';
+import { DomainException, Extension } from '@core/exceptions/filters/domain-exceptions';
 import { DomainExceptionCode } from '@core/exceptions/filters/domain-exception-codes';
 import BlogQueryRepository from '@modules/bloggers-platform/blogs/infrastructure/blogs.query-repository';
 import PostsQueryRepository from '@modules/bloggers-platform/posts/infrastructure/posts.query-repository';
@@ -23,7 +23,15 @@ export class DeletePostUseCase implements ICommandHandler<DeletePostCommand> {
   async execute(command: DeletePostCommand): Promise<void> {
     const { blogId, postId } = command;
     await this.blogQueryRepository.findOrNotFoundFail(blogId);
-    await this.postsQueryRepository.findOrNotFoundFail(postId);
+    const checkedPostId = await this.postsQueryRepository.findOrNotFoundFail(postId);
+
+    if (!checkedPostId.length) {
+      throw new DomainException({
+        code: DomainExceptionCode.NotFound,
+        message: 'Post not found',
+        extensions: [new Extension('Post with given id does not exist', 'id')],
+      });
+    }
 
     const deleted = await this.postsRepository.delete(postId);
     if (!deleted) {
